# Fix performance regression introduced in object deserialization