Special cases deserialization for the built-in object converter so that it avoids making expensive `ReadStack.Push()/Pop()` calls.

|                   Method |        Job |                                                                                                          Toolchain |     Mean |    Error |   StdDev |   Median |      Min |      Max | Ratio | MannWhitney(3%) |  Gen 0 |  Gen 1 | Allocated | Alloc Ratio |
|------------------------- |----------- |------------------------------------------------------------------------------------------------------------------- |---------:|---------:|---------:|---------:|---------:|---------:|------:|---------------- |-------:|-------:|----------:|------------:|
| DeserializeFromUtf8Bytes | Job-QKIRBU |  main | 43.18 us | 0.768 us | 0.681 us | 43.27 us | 41.77 us | 43.88 us |  1.00 |            Base | 4.0761 | 0.8492 |  40.36 KB |        1.00 |
| DeserializeFromUtf8Bytes | Job-XPCOIK | PR | 35.05 us | 0.570 us | 0.476 us | 35.04 us | 33.94 us | 35.92 us |  0.81 |          Faster | 4.0367 | 0.9744 |  39.93 KB |        0.99 |

Fixes performance regression introduced by https://github.com/dotnet/runtime/pull/65748. Fix #66341.