@DrewScoggins 2022-08-09
Improvement:
win-x64: https://github.com/dotnet/perf-autofiling-issues/issues/7178